move common struct to common_map.rs
