+++ FUSOU-APP/src-tauri/src/kcapi_common/mod.rs
@@ -1,3 +1,4 @@
 pub mod common_air;
 pub mod common_battle;
 pub mod common_midnight;
+pub mod common_map;
